requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
idna==2.9
joblib==0.15.1
nltk==3.5
oauthlib==3.1.0
python-dotenv==0.13.0
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
tqdm==4.46.0
urllib3==1.25.9
